sqlm_async/requirements.txt
aiosqlite==0.19.0
asyncpg==0.28.0
greenlet==2.0.2
markdown-it-py==3.0.0
mdurl==0.1.2
psycopg2-binary==2.9.6
pydantic==1.10.11
Pygments==2.15.1
python-dotenv==1.0.0
rich==13.4.2
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a34
sqlmodel==0.0.8
tqdm==4.65.0
typing_extensions==4.7.1
